backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
requests
python-dotenv
faiss-cpu
pdfplumber
python-docx
psutil
numpy